+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update to circle-stdlib v16/Circle Step 45.1.
 - Update ARM toolchains to 11.3.rel1.
 - Update to libmt32emu v2.7.0.
-- Update to FluidSynth v2.3.0.
+- Update to FluidSynth v2.3.1.
 - Update to inih r56.
 - Updater: `config.txt` is now replaced on update, however the `avoid_warnings` setting is preserved (the only setting that matters to most users).
 - Updater: improved error handling.